+++ lib/6to5/transformation/transform.js
@@ -94,9 +94,7 @@ _.each({
   _moduleFormatter:          require("./transformers/_module-formatter"),
 
   typeofSymbol:              require("./transformers/optional-typeof-symbol"),
-  externalRuntime:           require("./transformers/optional-external-runtime"),
-  coreAliasing:              require("./transformers/optional-core-aliasing"),
-  externalRegenerator:       require("./transformers/optional-external-regenerator-runtime"),
+  selfContained:             require("./transformers/optional-self-contained"),
   undefinedToVoid:           require("./transformers/optional-undefined-to-void"),
 
   // spec
