+++ lib/6to5/transformation/transformers/optional-self-contained.js
@@ -23,7 +23,9 @@ exports.optional = true;
 
 exports.ast = {
   enter: function (ast, file) {
+    file.opts.runtime = file.addImport("6to5-runtime/6to5", "to5Runtime").name;
     file._coreId = file.addImport("6to5-runtime/core-js", "core");
+    file._regeneratorId = file.addImport("6to5-runtime/regenerator", "regeneratorRuntime");
   },
 
   exit: function (ast, file) {
@@ -66,4 +68,11 @@ exports.ast = {
       }
     });
   }
+};
+
+exports.MemberExpression = function (node, parent, scope, context, file) {
+  var obj = node.object;
+  var prop = node.property;
+  if (!t.isReferenced(node, parent) || !t.isReferenced(obj, node)) return;
+  if (obj.name === "regeneratorRuntime") return t.memberExpression(file._regeneratorId, prop);
 };
