--- test/fixtures/transformation/optional-core-aliasing/es6-destructuring/actual.js
@@ -1,1 +1,0 @@
-var [a, [b], [c], d] = ["hello", [", ", "junk"], ["world"]];
