--- test/fixtures/transformation/optional-core-aliasing/es6-destructuring/expected.js
@@ -1,30 +1,0 @@
-"use strict";
-
-var _slicedToArray = function (arr, i) {
-  if (Array.isArray(arr)) {
-    return arr;
-  } else {
-    var _arr = [];
-
-    for (var _iterator = _core.$for.getIterator(arr), _step; !(_step = _iterator.next()).done;) {
-      _arr.push(_step.value);
-
-      if (i && _arr.length === i) break;
-    }
-
-    return _arr;
-  }
-};
-
-var _core = require("6to5-runtime/core-js");
-
-var _ref = ["hello", [", ", "junk"], ["world"]];
-
-var a = _ref[0];
-var _ref$1 = _slicedToArray(_ref[1], 1);
-
-var b = _ref$1[0];
-var _ref$2 = _slicedToArray(_ref[2], 1);
-
-var c = _ref$2[0];
-var d = _ref[3];
