--- test/fixtures/transformation/optional-core-aliasing/es6-spread/actual.js
@@ -1,1 +1,0 @@
-var lyrics = ["head", "and", "toes", ...parts];
