--- test/fixtures/transformation/optional-core-aliasing/es6-spread/expected.js
@@ -1,9 +1,0 @@
-"use strict";
-
-var _toArray = function (arr) {
-  return Array.isArray(arr) ? arr : _core.Array.from(arr);
-};
-
-var _core = require("6to5-runtime/core-js");
-
-var lyrics = ["head", "and", "toes"].concat(_toArray(parts));
