--- test/fixtures/transformation/optional-core-aliasing/options.json
@@ -1,4 +1,0 @@
-{
-  "optional": ["coreAliasing"],
-  "experimental": true
-}
