--- test/fixtures/transformation/optional-external-regenerator-runtime/options.json
@@ -1,3 +1,0 @@
-{
-  "optional": ["externalRegenerator"]
-}
