--- test/fixtures/transformation/optional-external-runtime/amd/expected.js
@@ -1,6 +1,0 @@
-define(["exports", "6to5-runtime/6to5", "someModule"], function (exports, _to5Runtime6to5, _someModule) {
-  "use strict";
-
-  var _to5Runtime = _to5Runtime6to5;
-  var foo = _to5Runtime.interopRequire(_someModule);
-});
