--- test/fixtures/transformation/optional-external-runtime/options.json
@@ -1,3 +1,0 @@
-{
-  "optional": ["externalRuntime"]
-}
