+++ test/fixtures/transformation/optional-self-contained/aliased-constructors/actual.js
