+++ test/fixtures/transformation/optional-self-contained/aliased-constructors/expected.js
@@ -1,6 +1,10 @@
 "use strict";
 
+var _to5Runtime = require("6to5-runtime/6to5");
+
 var _core = require("6to5-runtime/core-js");
+
+var _regeneratorRuntime = require("6to5-runtime/regenerator");
 
 obj.constructor === Object;
 obj.constructor === _core.Promise;
