+++ test/fixtures/transformation/optional-self-contained/es6-for-of/actual.js
