+++ test/fixtures/transformation/optional-self-contained/es6-for-of/expected.js
@@ -1,6 +1,10 @@
 "use strict";
 
+var _to5Runtime = require("6to5-runtime/6to5");
+
 var _core = require("6to5-runtime/core-js");
+
+var _regeneratorRuntime = require("6to5-runtime/regenerator");
 
 for (var _iterator = _core.$for.getIterator(arr), _step; !(_step = _iterator.next()).done;) {
   var i = _step.value;
