+++ test/fixtures/transformation/optional-self-contained/es7-array-comprehensions/actual.js
