+++ test/fixtures/transformation/optional-self-contained/options.json
@@ -1,0 +1,4 @@
+{
+  "optional": ["selfContained"],
+  "experimental": true
+}
