+++ test/fixtures/transformation/optional-self-contained/regenerator-runtime/actual.js
