+++ test/fixtures/transformation/optional-self-contained/regenerator-runtime/expected.js
@@ -1,5 +1,9 @@
 "use strict";
 
+var _to5Runtime = require("6to5-runtime/6to5");
+
+var _core = require("6to5-runtime/core-js");
+
 var _regeneratorRuntime = require("6to5-runtime/regenerator");
 
 void _regeneratorRuntime.mark(function callee$0$0() {
