+++ test/fixtures/transformation/optional-self-contained/runtime-amd/actual.js
