+++ test/fixtures/transformation/optional-self-contained/runtime-amd/expected.js
@@ -1,0 +1,8 @@
+define(["exports", "6to5-runtime/6to5", "6to5-runtime/core-js", "6to5-runtime/regenerator", "someModule"], function (exports, _to5Runtime6to5, _to5RuntimeCoreJs, _to5RuntimeRegenerator, _someModule) {
+  "use strict";
+
+  var _to5Runtime = _to5Runtime6to5;
+  var _core = _to5RuntimeCoreJs;
+  var _regeneratorRuntime = _to5RuntimeRegenerator;
+  var foo = _to5Runtime.interopRequire(_someModule);
+});
