+++ test/fixtures/transformation/optional-self-contained/runtime-amd/options.json
