+++ test/fixtures/transformation/optional-self-contained/runtime-commonjs/actual.js
