+++ test/fixtures/transformation/optional-self-contained/runtime-commonjs/expected.js
@@ -2,4 +2,8 @@
 
 var _to5Runtime = require("6to5-runtime/6to5");
 
+var _core = require("6to5-runtime/core-js");
+
+var _regeneratorRuntime = require("6to5-runtime/regenerator");
+
 var foo = _to5Runtime.interopRequireWildcard(require("someModule"));
